requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pyyaml==6.0.1
httpx==0.27.0
pytest==8.1.1
python-multipart==0.0.9